pip-requirements.txt
django==3.2
python-telegram-bot==13.14.1
qrcode==1.15.0